+++ pxl-javax/src/main/java/io/github/hclimkr/pxl/option/PxlImportSheetOption.java
@@ -1,6 +1,7 @@
 package io.github.hclimkr.pxl.option;
 
 import io.github.hclimkr.pxl.PxlConstants;
+import io.github.hclimkr.pxl.annotation.PxlSheet;
 import io.github.hclimkr.pxl.exception.PxlNullPointerException;
 import io.github.hclimkr.pxl.internal.constraint.Nullable;
 import io.github.hclimkr.pxl.internal.support.PxlAssertSupport;
@@ -32,6 +33,10 @@ public final class PxlImportSheetOption {
 
     /**
      * Specifies the name of the sheet on import.
+     * <p>
+     * Doubles as a content-i18n key, just like the {@link PxlSheet#name()} it overrides:
+     * when the workbook sets {@code importI18nBaseName}, the name is resolved through that bundle first and it is the
+     * translation that is matched. A name the bundle does not carry is used as it stands.
      */
     @Builder.Default
     private List<String> importSheetNames = null;